// src/models/VersionVersionMessage/types.ts
export interface VersionVersionMessageModelOptions {
  buildDate?: string;
  compiler?: string;
  extraBuildInfo?: string;
  gitCommit?: string;
  gitTag?: string;
  gitTreeState?: string;
  goVersion?: string;
  helmVersion?: string;
  jsonnetVersion?: string;
  kubectlVersion?: string;
  kustomizeVersion?: string;
  platform?: string;
  version?: string;
}